Streamline developer experience
Unnecessary tick has been removed, it was not useful as the queue of action would ge the priority, and the select is made in a way that it waits for any chan output

Improve Context depending on usage. A deliberated choice has been made to outsource context safety of the action execution on the consumer side. This will be revisited at a later stage